+++ src/Symfony/Component/Messenger/Tests/Stamp/TransportNamesStampTest.php
@@ -17,6 +17,8 @@ use Symfony\Component\Messenger\Stamp\TransportNamesStamp;
 use Symfony\Component\Messenger\Transport\Serialization\Serializer;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainNormalizer;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Serializer as SymfonySerializer;
 
@@ -38,10 +40,12 @@ class TransportNamesStampTest extends TestCase
     {
         $stamp = new TransportNamesStamp(['foo']);
         $serializer = new Serializer(
-            new SymfonySerializer([
-                new ArrayDenormalizer(),
-                new ObjectNormalizer(),
-            ], [new JsonEncoder()])
+            new SymfonySerializer(
+                [],
+                [new JsonEncoder()],
+                new ChainNormalizer([new ObjectNormalizer()]),
+                new ChainDenormalizer([new ArrayDenormalizer(), new ObjectNormalizer()]),
+            )
         );
 
         $deserializedEnvelope = $serializer->decode($serializer->encode(new Envelope(new \stdClass(), [$stamp])));