+++ src/Symfony/Component/Mime/Tests/EmailTest.php
@@ -26,6 +26,8 @@ use Symfony\Component\Mime\Test\Constraint\EmailHeaderSame;
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ChainNormalizer;
 use Symfony\Component\Serializer\Normalizer\MimeMessageNormalizer;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
 use Symfony\Component\Serializer\Normalizer\PropertyNormalizer;
@@ -570,12 +572,21 @@ EOF;
 
         $extractor = new PhpDocExtractor();
         $propertyNormalizer = new PropertyNormalizer(null, null, $extractor);
-        $serializer = new Serializer([
-            new ArrayDenormalizer(),
-            new MimeMessageNormalizer($propertyNormalizer),
-            new ObjectNormalizer(null, null, null, $extractor),
-            $propertyNormalizer,
-        ], [new JsonEncoder()]);
+        $serializer = new Serializer(
+            [],
+            [new JsonEncoder()],
+            new ChainNormalizer([
+                new MimeMessageNormalizer($propertyNormalizer),
+                new ObjectNormalizer(null, null, null, $extractor),
+                $propertyNormalizer,
+            ]),
+            new ChainDenormalizer([
+                new ArrayDenormalizer(),
+                new MimeMessageNormalizer($propertyNormalizer),
+                new ObjectNormalizer(null, null, null, $extractor),
+                $propertyNormalizer,
+            ])
+        );
 
         $serialized = $serializer->serialize($e, 'json', [ObjectNormalizer::IGNORED_ATTRIBUTES => ['cachedBody']]);
         $this->assertSame($expectedJson, json_encode(json_decode($serialized), \JSON_PRETTY_PRINT | \JSON_UNESCAPED_SLASHES));